:sparkles: :: 현재 외출중인 학생을 검색할 수 있는 권한을 학생에게도 주기
