+++ src/main/kotlin/com/project/goms/domain/studentCouncil/presentation/StudentCouncilController.kt
@@ -1,10 +1,10 @@
 package com.project.goms.domain.studentCouncil.presentation
 
 import com.project.goms.domain.account.entity.Authority
+import com.project.goms.domain.outing.usecase.SearchOutingUseCase
 import com.project.goms.domain.studentCouncil.common.util.StudentCouncilConverter
 import com.project.goms.domain.studentCouncil.presentation.data.request.GrantAuthorityRequest
 import com.project.goms.domain.studentCouncil.presentation.data.response.AllAccountResponse
-import com.project.goms.domain.studentCouncil.presentation.data.response.SearchOutingResponse
 import com.project.goms.domain.studentCouncil.usecase.*
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
@@ -68,11 +68,5 @@ class StudentCouncilController(
     fun deleteOuting(@PathVariable accountIdx: UUID): ResponseEntity<Void> =
         deleteOutingUseCase.execute(accountIdx)
             .let { ResponseEntity.status(HttpStatus.RESET_CONTENT).build() }
-
-    @GetMapping("search/outing")
-    fun searchOuting(@RequestParam name: String): ResponseEntity<List<SearchOutingResponse>> =
-        searchOutingUseCase.execute(name)
-            .let { studentCouncilConverter.toSearchOutingResponse(it) }
-            .let { ResponseEntity.ok(it) }
 
 }
